String fix for chkpt name
styling

+++ src/components/lib/crt/crt.h
@@ -164,8 +164,8 @@ int crt_comp_create(struct crt_comp *c, char *name, compid_t id, void *elf_hdr, 
 int crt_comp_create_with(struct crt_comp *c, char *name, compid_t id, struct crt_comp_resources *resources);
 
 int crt_comp_create_from(struct crt_comp *c, char *name, compid_t id, struct crt_chkpt *chkpt);
-int crt_ncomp();
-int crt_nchkpt();
+unsigned long crt_ncomp();
+unsigned long crt_nchkpt();
 
 int crt_comp_alias_in(struct crt_comp *c, struct crt_comp *c_in, struct crt_comp_resources *res, crt_comp_alias_t flags);
 void crt_comp_captbl_frontier_update(struct crt_comp *c, capid_t capid);
